Organize blockhash as a module
This makes blockhash install as a module which makes it possible to be
packaged in bazel rules, but is probably also slightly more useful in
general.

This also updates to use setuptools and converts the script to an
entry_point which will make it more broadly supported on different
platforms.